reverse the dependency between pattern and gflow
