+++ graphix/__init__.py
@@ -1,5 +1,5 @@
 from graphix.transpiler import Circuit
-from graphix.gflow import generate_from_graph
+from graphix.generator import generate_from_graph
 from graphix.pattern import Pattern
 from graphix.sim.statevec import Statevec
 from graphix.graphsim import GraphState
