+++ graphix/gflow.py
@@ -15,105 +15,6 @@ import networkx as nx
 import numpy as np
 import z3
 import copy
-from graphix.pattern import Pattern
-
-
-def generate_from_graph(graph, angles, inputs, outputs, timeout=100):
-    r"""Generate the measurement pattern from open graph and measurement angles.
-
-    This function takes an open graph G = (nodes, edges, input, outputs),
-    specified by networks.Graph and two lists specifying input and output nodes.
-    Currently we support XY-plane measurements.
-
-    Searches for the flow in the open graph using :func:`flow` and if found,
-    construct the measurement pattern according to the theorem 1 of [NJP 9, 250 (2007)].
-
-    Then, if no flow was found, searches for gflow using :func:`gflow`,
-    from which measurement pattern can be constructed from theorem 2 of [NJP 9, 250 (2007)].
-
-    The constructed measurement pattern deterministically realize the unitary embedding
-
-    .. math::
-
-        U = \left( \prod_i \langle +_{\alpha_i} |_i \right) E_G N_{I^C},
-
-    where the measurements (bras) with always :math:`\langle+|` bases determined by the measurement
-    angles :math:`\alpha_i` are applied to the measuring nodes,
-    i.e. the randomness of the measurement is eliminated by the added byproduct commands.
-
-    .. seealso:: :func:`flow` :func:`gflow` :class:`graphix.pattern.Pattern`
-
-    Parameters
-    ----------
-    graph : networkx.Graph
-        graph on which MBQC should be performed
-    angles : dict
-        measurement angles for each nodes on the graph (unit of pi), except output nodes
-    inputs : list
-        list of node indices for input nodes
-    outputs : list
-        list of node indices for output nodes
-    timeout : int
-        optional argument for flow and gflow search depth
-
-    Returns
-    -------
-    pattern : graphix.pattern.Pattern object
-        constructed pattern.
-    """
-    assert len(inputs) == len(outputs)
-    measuring_nodes = list(set(graph.nodes) - set(outputs) - set(inputs))
-
-    # search for flow first
-    f, l_k = flow(graph, set(inputs), set(outputs), timeout=timeout)
-    if f:
-        # flow found
-        depth, layers = get_layers(l_k)
-        pattern = Pattern(len(inputs))
-        pattern.seq = [["N", i] for i in inputs]
-        for i in set(graph.nodes) - set(inputs):
-            pattern.seq.append(["N", i])
-        for e in graph.edges:
-            pattern.seq.append(["E", e])
-        measured = []
-        for i in range(depth, 0, -1):  # i from depth, depth-1, ... 1
-            for j in layers[i]:
-                measured.append(j)
-                pattern.seq.append(["M", j, "XY", angles[j], [], []])
-                for k in set(graph.neighbors(f[j])) - set([j]):
-                    if k not in measured:
-                        pattern.seq.append(["Z", k, [j]])
-                pattern.seq.append(["X", f[j], [j]])
-        pattern.output_nodes = outputs
-        pattern.Nnode = len(graph.nodes)
-    else:
-        # no flow found - we try gflow
-        g, l_k = gflow(graph, set(inputs), set(outputs), timeout=timeout)
-        if g:
-            # gflow found
-            depth, layers = get_layers(l_k)
-            pattern = Pattern(len(inputs))
-            pattern.seq = [["N", i] for i in inputs]
-            for i in set(graph.nodes) - set(inputs):
-                pattern.seq.append(["N", i])
-            for e in graph.edges:
-                pattern.seq.append(["E", e])
-            remaining = set(measuring_nodes)
-            for i in range(depth, 0, -1):  # i from depth, depth-1, ... 1
-                for j in layers[i]:
-                    pattern.seq.append(["M", j, "XY", angles[j], [], []])
-                    remaining = remaining - set([j])
-                    odd_neighbors = find_odd_neighbor(graph, remaining, set(g[j]))
-                    for k in odd_neighbors:
-                        pattern.seq.append(["Z", k, [j]])
-                    for k in set(g[j]) - set([j]):
-                        pattern.seq.append(["X", k, [j]])
-            pattern.output_nodes = outputs
-            pattern.Nnode = len(graph.nodes)
-        else:
-            raise ValueError("no flow or gflow found")
-
-    return pattern
 
 
 def solvebool(A, b):
@@ -172,7 +73,7 @@ def solvebool(A, b):
         return False
 
 
-def gflow(g, v_in, v_out, meas_plane=None, timeout=100):
+def gflow(g, v_in, v_out, meas_plane=None, timeout=1000):
     """Optimum generalized flow finding algorithm
 
     For open graph g with input and output, this returns optimum gflow.
@@ -286,7 +187,6 @@ def gflowaux(v, gamma, index_list, v_in, v_out, g, l_k, k, meas_plane):
 
     # if index_0 or index_1 is blank, skip the calculation below
     if len(index_0) * len(index_1):
-
         gamma_sub = gamma[index_0, index_1]
 
         for u in iter(v_rem):
@@ -322,7 +222,7 @@ def gflowaux(v, gamma, index_list, v_in, v_out, g, l_k, k, meas_plane):
     return v_out | c_set, g, l_k, finished, exist
 
 
-def flow(g, v_in, v_out, meas_plane=None, timeout=100):
+def flow(g, v_in, v_out, meas_plane=None, timeout=10000):
     """Causal flow finding algorithm
 
     For open graph g with input and output, this returns causal flow.
@@ -561,67 +461,3 @@ def get_layers(l_k):
     for i in l_k.keys():
         layers[l_k[i]].append(i)
     return d, layers
-
-
-def get_meas_plane(pattern):
-    """get measurement plane from the pattern.
-    Parameter
-    -------
-    pattern: graphix.pattern.Pattern object
-
-    Return
-    -------
-    meas_plane: dict of str
-        list of strs representing measurement plane for each node.
-    """
-    meas_plane = dict()
-    XYtoXZ_list = [
-        6,
-        8,
-        11,
-        12,
-        20,
-        21,
-        22,
-        23,
-    ]  # clifford indices that change meas plane from xy to xz
-    for cmd in pattern.seq:
-        if cmd[0] == "M":
-            if len(cmd) == 7:
-                if cmd[6] in XYtoXZ_list:
-                    meas_plane[cmd[1]] = "XZ"
-                else:
-                    meas_plane[cmd[1]] = cmd[2]
-            else:
-                meas_plane[cmd[1]] = cmd[2]
-    return meas_plane
-
-
-def get_measurement_order_from_gflow(pattern):
-    """Returns a list containing the node indices,
-    in the order of measurements which can be performed with minimum depth.
-    Input must be a simple connected graph for the gflow-finding algorithm to work.
-
-    Returns
-    -------
-    meas_order : list
-        list of node indices in the order of measurements
-    """
-    nodes, edges = pattern.get_graph()
-    G = nx.Graph()
-    G.add_nodes_from(nodes)
-    G.add_edges_from(edges)
-    isolated = list(nx.isolates(G))
-    if isolated:
-        raise ValueError("The input graph must be connected")
-    meas_plane = get_meas_plane(pattern)
-    g, l_k = gflow(G, set(), set(pattern.output_nodes), meas_plane=meas_plane)
-    if not g:
-        raise ValueError("No gflow found")
-    k, layers = get_layers(l_k)
-    meas_order = []
-    while k > 0:
-        for node in layers[k]:
-            meas_order.append(node)
-        k -= 1
-    return meas_order
