+++ graphix/pattern.py
@@ -5,6 +5,7 @@ import numpy as np
 import networkx as nx
 from graphix.simulator import PatternSimulator
 from graphix.graphsim import GraphState
+from graphix.gflow import flow, gflow, get_layers
 from graphix.clifford import CLIFFORD_CONJ, CLIFFORD_TO_QASM3, CLIFFORD_MUL
 from copy import deepcopy
 
@@ -542,7 +543,7 @@ class Pattern:
 
         Returns
         -------
-        dependency : dict of sets
+        dependency : dict of set
             index is node number. all nodes in the each set must be measured before measuring
         """
         nodes, _ = self.get_graph()
@@ -557,18 +558,18 @@ class Pattern:
         return dependency
 
     def update_dependency(self, measured, dependency):
-        """Update dependency function. remove measured nodes from dependency.
+        """Remove measured nodes from the 'dependency'.
 
         Parameters
         ----------
-        measured: set
+        measured: set of int
             measured nodes.
-        dependency: dict of sets
+        dependency: dict of set
             which is produced by `_get_dependency`
 
         Returns
         --------
-        dependency: dict of sets
+        dependency: dict of set
             updated dependency information
         """
         for i in dependency.keys():
@@ -584,7 +585,7 @@ class Pattern:
         -------
         depth : int
             depth of graph
-        layers : dict of sets
+        layers : dict of set
             nodes grouped by layer index(k)
         """
         dependency = self._get_dependency()
@@ -609,12 +610,11 @@ class Pattern:
         return depth, l_k
 
     def _measurement_order_depth(self):
-        """Obtain the measurement order which reduces the depth of
-        pattern execution.
+        """Obtain a measurement order which reduces the depth of a pattern.
 
         Returns
         -------
-        meas_order: list of ints
+        meas_order: list of int
             optimal measurement order for parallel computing
         """
         d, l_k = self.get_layers()
@@ -629,7 +629,7 @@ class Pattern:
 
         Returns
         -------
-        connected: set of taples
+        connected: set of taple
                 set of connected edges
         """
         connected = set()
@@ -641,13 +641,13 @@ class Pattern:
         return connected
 
     def _measurement_order_space(self):
-        """Determine the optimal measurement order
-         which minimize the `max_space` of the pattern.
+        """Determine a sub-optimal measurement order
+         which heuristically minimizes the `max_space` of a pattern.
 
         Returns
         -------
-        meas_order: list of ints
-            optimal measurement order for classical simulation
+        meas_order: list of int
+            sub-optimal measurement order for classical simulation
         """
         nodes, edges = self.get_graph()
         nodes = set(nodes)
@@ -674,48 +674,107 @@ class Pattern:
             edges -= removable_edges
         return meas_order
 
+    def get_measurement_order_from_gflow(self):
+        """Returns a list containing the node indices,
+        in the order of measurements which can be performed with minimum depth.
+        Input must be a simple connected graph for the gflow-finding algorithm to work.
+
+        Returns
+        -------
+        meas_order : list of int
+            measurement order
+        """
+        nodes, edges = self.get_graph()
+        G = nx.Graph()
+        G.add_nodes_from(nodes)
+        G.add_edges_from(edges)
+        isolated = list(nx.isolates(G))
+        if isolated:
+            raise ValueError("The input graph must be connected")
+        meas_plane = self.get_meas_plane()
+        g, l_k = gflow(G, set(), set(self.output_nodes), meas_plane=meas_plane)
+        if not g:
+            raise ValueError("No gflow found")
+        k, layers = get_layers(l_k)
+        meas_order = []
+        while k > 0:
+            for node in layers[k]:
+                meas_order.append(node)
+            k -= 1
+        return meas_order
+
     def sort_measurement_commands(self, meas_order):
         """Convert measurement order to sequence of measurement commands
 
         Parameters
-        --------
-        meas_order: list of ints
+        ----------
+        meas_order: list of int
             optimal measurement order.
 
         Returns
-        --------
-        meas_flow: list of commands
+        -------
+        meas_cmds: list of command
             sorted measurement commands
         """
-        meas_flow = []
+        meas_cmds = []
         for i in meas_order:
             target = 0
             while True:
                 if (self.seq[target][0] == "M") & (self.seq[target][1] == i):
-                    meas_flow.append(self.seq[target])
+                    meas_cmds.append(self.seq[target])
                     break
                 target += 1
-        return meas_flow
+        return meas_cmds
 
-    def get_measurement_order(self):
-        """Returns the list containing the node indices,
+    def get_measurement_commnads(self):
+        """Returns the list containing the measurement commands,
         in the order of measurements
 
         Returns
         -------
-        meas_flow : list
-            list of node indices in the order of meaurements
+        meas_cmds : list
+            list of measurement commands in the order of meaurements
         """
         if not self.is_standard():
             self.standardize()
-        meas_flow = []
+        meas_cmds = []
         ind = self._find_op_to_be_moved("M")
         if ind == "end":
             return []
         while self.seq[ind][0] == "M":
-            meas_flow.append(self.seq[ind])
+            meas_cmds.append(self.seq[ind])
             ind += 1
-        return meas_flow
+        return meas_cmds
+
+    def get_meas_plane(self):
+        """get measurement plane from the pattern.
+
+        Returns
+        -------
+        meas_plane: dict of str
+            list of str representing measurement plane for each node.
+        """
+        meas_plane = dict()
+        XYtoXZ_list = [
+            6,
+            8,
+            11,
+            12,
+            20,
+            21,
+            22,
+            23,
+        ]  # clifford indices that change meas plane from xy to xz
+        for cmd in self.seq:
+            if cmd[0] == "M":
+                if len(cmd) == 7:
+                    if cmd[6] in XYtoXZ_list:
+                        meas_plane[cmd[1]] = "XZ"
+                    else:
+                        meas_plane[cmd[1]] = cmd[2]
+                else:
+                    meas_plane[cmd[1]] = cmd[2]
+        return meas_plane
 
     def get_graph(self):
         """returns the list of nodes and edges from the command sequence,
@@ -741,7 +800,7 @@ class Pattern:
 
         Returns
         -------
-        set of int :
+        isolated_nodes : set of int
             set of the isolated nodes
         """
         nodes, edges = self.get_graph()
@@ -749,7 +808,8 @@ class Pattern:
         connected_node_set = set()
         for edge in edges:
             connected_node_set |= set(edge)
-        return node_set - connected_node_set
+        isolated_nodes = node_set - connected_node_set
+        return isolated_nodes
 
     def get_vops(self, conj=False, include_identity=False):
         """Get local-Clifford decorations from measurement or Clifford commands.
@@ -864,7 +924,7 @@ class Pattern:
 
         Parameters
         ----------
-        meas_commands : list of commands
+        meas_commands : list of command
             list of measurement ('M') commands
         """
         prepared = []
@@ -1115,7 +1175,7 @@ def pauli_nodes(pattern):
     """
     if not pattern.is_standard():
         pattern.standardize()
-    m_commands = pattern.get_measurement_order()
+    m_commands = pattern.get_measurement_commands()
     pauli_node = []
     # Nodes that are non-Pauli measured, or pauli measured but depends on pauli measurement
     non_pauli_node = []
